requirements.txt
# Python standard library only
python-dotenv
requests